Handle Statistics and SparseArrays as extensions
